before Full redo of Debug

+++ Debug.h
@@ -6,7 +6,7 @@ using std::cerr;
 using std::endl;
 using std::vector;
 typedef vector< vector<float> > vec;
-#define DB_FLAG vector<int>({0,12})
+#define DB_FLAG vector<float>({0,12})
 #define DBA(x)do{cerr << x << endl;}while(0)
 #define DBVA(x) do{for(auto h : x) { cerr << h << "  ";}cerr << endl; } while(0)
 //#define DB(x, level) if(DB_FLAG == level)do{cerr << x << endl;}while(0)
